AVF: Update the HAL Implementation
Currently the VM HAL has hardcoded the DTBO idx, the suffix for the QTVM
DTBO partition and the disablement of earlyvirtmgr. This change is to
update the DTBO idx to be dynamically read from a system property,
the slot suffix is read from a system property and in the json
config the early virtmgr enablement is disabled.

Change-Id: I7f2d6e1d8172ab4d1b69c5965fe8fd444efaf657