AVF HAL: Update Default Config and Refactor
Update the total memory default to be CMA + shared mem. Also now CMA
size is configurable, with the CMA size being capped at the max size of
the file. Shared mem is defaulted to 2mb for both VMs and vCPUs is
defaulted to 4. Added the Persist image to the json config. Refactored
the code folder structure.

Change-Id: Iedbe745fbcda31f30740a8de32dbf590a7226278